Add pre-commit configuration
As we're fighting the CI a lot for simple validation, this commit adds
a pre-commit integration that'll perform linting and style checks
locally before GitHub actions gets a chance to make us look silly
in the automated messages channel.

At present this just covers the following lints:

- Trailing whitespace
- Blank lines at end of file
- Line ending normalization
- Basic XML and YAML validation (without schemas)
- Removal of any UTF8 BOMs
- Running Luacheck on all scripts
- Running Stylua for style enforcement

The combination of all the above _largely_ obsoletes the editorconfig
checks, which I'd rather move toward getting rid of as they're not
aware of language-specific constructs; for example we can't run
editorconfig-checker on our localization scripts because they'll
complain about trailing whitespace for multi-line strings.

What hasn't been added yet is XML schema validation, and replacing
the existing check steps in the action workflows to just go through
these steps so there's consistency both at pre-commit level _and_
at a PR level.

One additional change with all this - we're now normalizing line endings
to CRLF. Stylua wants us to pick between one or the other, and other
parts of the pre-commit checks _also_ want that. As the packager is
always shipping CRLF line endings anyway, it makes sense to just
align with that - which potentially means we can tell it to not do
its own normalization and significantly speed up its own builds.